chore(node-api): move FullNodeComponents from builder
