+++ crates/node-api/Cargo.toml
@@ -16,3 +16,7 @@ reth-evm.workspace = true
 reth-provider.workspace = true
 reth-db.workspace = true
 reth-engine-primitives.workspace = true
+reth-transaction-pool.workspace = true
+reth-network.workspace = true
+reth-payload-builder.workspace = true
+reth-tasks.workspace = true
